Notification refactors: t->type, MEMBERSHIP_MS_UPDATE, MEMBERSHIP_LS_UPDATE
