+++ src/__tests__/mocks/message.MEMBERSHIP_MS_UPDATE.UNBOUNDED.1457552640000.json
@@ -1,0 +1,4 @@
+{
+  "type": "message",
+  "data": "{\"data\":\"{\\\"type\\\":\\\"MEMBERSHIP_MS_UPDATE\\\",\\\"cn\\\":1457552640000,\\\"u\\\": 0,\\\"h\\\":0}\"}"
+}
