--- src/__tests__/mocks/message.MY_LARGE_SEGMENTS_UPDATE.UNBOUNDED.DELAY.1457552650000.json
@@ -1,4 +1,0 @@
-{
-  "type": "message",
-  "data": "{\"data\":\"{\\\"t\\\":\\\"MY_LARGE_SEGMENTS_UPDATE\\\",\\\"cn\\\":1457552650000,\\\"n\\\":[],\\\"c\\\": 0,\\\"u\\\": 0,\\\"d\\\":\\\"\\\",\\\"i\\\":300,\\\"h\\\":1,\\\"s\\\":0}\"}"
-}
