--- src/__tests__/mocks/message.MY_SEGMENTS_UPDATE_V3.BOUNDED.GZIP.1457552651000.json
@@ -1,4 +1,0 @@
-{
-  "type": "message",
-  "data": "{\"data\":\"{\\\"t\\\":\\\"MY_SEGMENTS_UPDATE_V3\\\",\\\"cn\\\":1457552651000,\\\"n\\\":[],\\\"c\\\": 1,\\\"u\\\": 1,\\\"d\\\":\\\"H4sIAAAAAAAA/2JABxzYeIxQLguYFIBLN8Bl4EABjc+EzOnAsA4QAAD//8YBvWeAAAAA\\\",\\\"h\\\":0}\"}"
-}
