--- src/__tests__/mocks/message.MY_SEGMENTS_UPDATE_V3.BOUNDED.ZLIB.1457552651000.json
@@ -1,4 +1,0 @@
-{
-  "type": "message",
-  "data": "{\"data\":\"{\\\"t\\\":\\\"MY_SEGMENTS_UPDATE_V3\\\",\\\"cn\\\":1457552651000,\\\"n\\\":[],\\\"c\\\": 2,\\\"u\\\": 1,\\\"d\\\":\\\"eJxiGAX4AMdAO2AU4AeMA+2AAQACA+0AuoORGMvDBDANtAPoDBQG2gGDGQz16pRloB0wCkbBKBgFo4As0EBYyZCqoojwDwEACAAA//+W/QFR\\\",\\\"h\\\":0}\"}"
-}
