--- src/__tests__/mocks/message.MY_SEGMENTS_UPDATE_V3.UNBOUNDED.1457552640000.json
@@ -1,4 +1,0 @@
-{
-  "type": "message",
-  "data": "{\"data\":\"{\\\"t\\\":\\\"MY_SEGMENTS_UPDATE_V3\\\",\\\"cn\\\":1457552640000,\\\"u\\\": 0,\\\"h\\\":0}\"}"
-}
